# back/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /store/{storeId}/inventory-items/products/{productId}:
    get:
      description: Retorna informações de estoque para um produto específico
      operationId: InventoryItemsController_getInventoryItemByProduct
      parameters:
        - name: storeId
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
        - name: productId
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      responses:
        '200':
          description: Item de inventário encontrado
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para acessar este recurso
        '404':
          description: Loja ou produto não encontrado no inventário
      security: *ref_4
      summary: Obter item do inventário por produto
      tags: *ref_5
    patch:
      description: Atualiza as informações de estoque para um produto específico
      operationId: InventoryItemsController_updateInventoryItem
      parameters:
        - name: storeId
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
        - name: productId
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      requestBody:
        required: true
        description: Dados para atualização do item de inventário
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryItemDto'
            examples:
              updateQuantity:
                summary: Atualizar quantidade
                description: Atualiza a quantidade em estoque para o produto
                value:
                  quantity: 75
      responses:
        '200':
          description: Item de inventário atualizado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar itens do inventário
        '404':
          description: Loja ou produto não encontrado
      security: *ref_4
      summary: Atualizar item do inventário
      tags: *ref_5
    delete:
      description: Remove um produto do estoque da loja
      operationId: InventoryItemsController_removeProduct
      parameters:
        - name: storeId
          required: true
          in: path
          description: ID da loja
          schema:
            type: number
        - name: productId
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      responses:
        '200':
          description: Produto removido do estoque com sucesso
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para remover produtos do estoque
        '404':
          description: Loja ou produto não encontrado
      security: *ref_4
      summary: Remover produto do estoque
      tags: *ref_5
...
  /products:
    post:
      description: Cria um novo produto para a loja do produtor autenticado
      operationId: ProductsController_create
      parameters: []
      requestBody:
        required: true
        description: Dados do produto a ser criado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
            examples:
              vegetable:
                summary: Produto vegetal
                description: Exemplo de produto da categoria vegetais
                value:
                  name: Alface Crespa
                  description: Alface crespa fresca, cultivada sem agrotóxicos
                  price: 3.5
                  imageUrl: https://exemplo.com/imagens/alface.jpg
                  rating: 4.8
                  category: Vegetais
                  storeId: 1
              fruit:
                summary: Produto frutas
                description: Exemplo de produto da categoria frutas
                value:
                  name: Morango Orgânico
                  description: Morangos orgânicos cultivados na região serrana
                  price: 12.9
                  imageUrl: https://exemplo.com/imagens/morango.jpg
                  rating: 4.5
                  category: Frutas
                  storeId: 1
      responses:
        '201':
          description: Produto criado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para criar produtos
      security:
        - JWT-auth: []
      summary: Criar um novo produto
      tags: &ref_6
        - Products
    get:
      description: Retorna todos os produtos ou filtra por nome
      operationId: ProductsController_findAllOrFindByName
      parameters:
        - name: name
          required: false
          in: query
          description: Nome do produto para filtrar
          schema:
            type: string
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
      summary: Listar produtos
      tags: *ref_6
  /products/{id}:
    get:
      description: Retorna um produto específico pelo seu ID
      operationId: ProductsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      responses:
        '200':
          description: Produto encontrado
        '404':
          description: Produto não encontrado
      summary: Buscar produto por ID
      tags: *ref_6
    patch:
      description: Atualiza os dados de um produto existente
      operationId: ProductsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      requestBody:
        required: true
        description: Dados para atualização do produto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
            examples:
              updatePrice:
                summary: Atualizar preço
                description: Atualiza apenas o preço do produto
                value:
                  price: 4.25
              updateDetails:
                summary: Atualizar detalhes
                description: Atualiza descrição e imagem do produto
                value:
                  description: Nova descrição atualizada do produto
                  imageUrl: https://exemplo.com/imagens/produto-atualizado.jpg
              updateMultiple:
                summary: Atualização completa
                description: Atualiza múltiplos campos do produto
                value:
                  name: Nome atualizado
                  description: Descrição atualizada
                  price: 5.99
                  rating: 4.9
                  category: Nova Categoria
      responses:
        '200':
          description: Produto atualizado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para atualizar este produto
        '404':
          description: Produto não encontrado
      security:
        - JWT-auth: []
      summary: Atualizar produto
      tags: *ref_6
    delete:
      description: Remove um produto existente
      operationId: ProductsController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: ID do produto
          schema:
            type: number
      responses:
        '200':
          description: Produto removido com sucesso
        '401':
          description: Usuário não autenticado
        '403':
          description: Usuário não tem permissão para remover este produto
        '404':
          description: Produto não encontrado
      security:
        - JWT-auth: []
      summary: Remover produto
      tags: *ref_6
...
info:
  title: Ze da Horta API
  description: API documentation for the NestJS application
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Insira seu token JWT
      in: header
  schemas:
...
    UpdateInventoryItemDto:
      type: object
      properties: {}
    CreateProductDto:
      type: object
      properties: {}
    UpdateProductDto:
      type: object
      properties: {}
...